refactor: 팀 컨벤션에 맞게 components 폴더 생성
