+++ src/app/(auth)/login/email/page.tsx
@@ -8,8 +8,8 @@ import { useRouter } from 'next/navigation';
 import useAuthStore from '@/store/authStore';
 import axiosInstance from '@/app/api/axiosInstance';
 import { AxiosError } from 'axios';
-import { EyeSlashFilledIcon } from './EyeSlashFilledIcon';
-import { EyeFilledIcon } from './EyeFilledIcon';
+import { EyeSlashFilledIcon } from '../components/EyeSlashFilledIcon';
+import { EyeFilledIcon } from '../components/EyeFilledIcon';
 
 const EmailLogin = () => {
   const router = useRouter();
