+++ src/app/(auth)/user-info/components/UserPosition.tsx
