+++ src/app/map/components/CourtDetail.tsx
